+++ DataAccessTier/IRepository/JobManagement/JobInformation/ISubmittedTaskInfoRepository.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using HEAppE.DomainObjects.JobManagement.JobInformation;
 
 namespace HEAppE.DataAccessTier.IRepository.JobManagement.JobInformation;
@@ -7,4 +7,7 @@ public interface ISubmittedTaskInfoRepository : IRepository<SubmittedTaskInfo>
 {
     IEnumerable<SubmittedTaskInfo> GetAllUnFinished();
     IEnumerable<SubmittedTaskInfo> GetAllFinished();
+    IEnumerable<SubmittedTaskInfo> GetFinishedByIds(IEnumerable<long> ids);
+    SubmittedTaskInfo GetByIdWithJobSpecification(long id);
+    SubmittedTaskInfo GetByIdWithProject(long id);
 }