+++ DataAccessTier/MiddlewareContext.cs
@@ -23,12 +23,20 @@ using Microsoft.Data.SqlClient;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.EntityFrameworkCore.Storage;
+using Microsoft.Extensions.Logging;
 
 namespace HEAppE.DataAccessTier;
 
-internal class MiddlewareContext : DbContext
+public class MiddlewareContext : DbContext
 {
     #region Constructors
+
+    private readonly ILogger _logger;
+
+    public MiddlewareContext(ILogger logger) : this()
+    {
+        _logger = logger;
+    }
 
     public MiddlewareContext()
     {
@@ -266,8 +274,74 @@ internal class MiddlewareContext : DbContext
                 parameter);
 
             modelBuilder.Entity(entityType.ClrType).HasQueryFilter(filter);
-            modelBuilder.Entity(entityType.ClrType).HasIndex([nameof(ISoftDeletableEntity.IsDeleted)]);
+            
+            modelBuilder.Entity(entityType.ClrType)
+                .HasIndex([nameof(ISoftDeletableEntity.IsDeleted)])
+                .HasFilter("[IsDeleted] = 0");
         }
+
+        modelBuilder.Entity<SubmittedTaskInfo>()
+            .HasIndex("SubmittedJobInfoId", nameof(SubmittedTaskInfo.State))
+            .IncludeProperties("SpecificationId", "NodeTypeId", "ProjectId");
+
+        modelBuilder.Entity<SubmittedTaskInfo>()
+            .HasIndex("SubmittedJobInfoId")
+            .IncludeProperties(nameof(SubmittedTaskInfo.State), "NodeTypeId", "SpecificationId");
+        
+        modelBuilder.Entity<SubmittedTaskInfo>()
+            .HasIndex(t => t.State)
+            .HasFilter("[State] >= 16")
+            .IncludeProperties("ProjectId", "SpecificationId", "SubmittedJobInfoId");
+        
+        modelBuilder.Entity<SubmittedTaskInfo>()
+            .HasIndex("SubmittedJobInfoId", nameof(SubmittedTaskInfo.State))
+            .HasFilter("[State] > 1 AND [State] < 16")
+            .IncludeProperties("NodeTypeId", "SpecificationId", "ProjectId");
+
+        modelBuilder.Entity<SubmittedJobInfo>()
+            .HasIndex("SpecificationId", "ProjectId")
+            .IncludeProperties(nameof(SubmittedJobInfo.State), "SubmitterId");
+
+        modelBuilder.Entity<TaskSpecification>()
+            .HasIndex("JobSpecificationId")
+            .IncludeProperties("CommandTemplateId", "ClusterNodeTypeId");
+
+        modelBuilder.Entity<SubmittedTaskAllocationNodeInfo>()
+            .HasIndex(nameof(SubmittedTaskAllocationNodeInfo.SubmittedTaskInfoId));
+
+        modelBuilder.Entity<SubmittedJobInfo>()
+            .HasIndex("ProjectId", nameof(SubmittedJobInfo.StartTime), nameof(SubmittedJobInfo.EndTime));
+
+        modelBuilder.Entity<SubmittedJobInfo>()
+            .HasIndex("SubmitterId")
+            .IncludeProperties(nameof(SubmittedJobInfo.State));
+
+        modelBuilder.Entity<SubmittedTaskInfo>()
+            .HasIndex(t => t.ScheduledJobId)
+            .HasFilter("[ScheduledJobId] IS NOT NULL");
+
+        modelBuilder.Entity<ClusterProjectCredentialCheckLog>()
+            .HasIndex("ClusterAuthenticationCredentialsId");
+
+        modelBuilder.Entity<ClusterNodeTypeAggregationAccounting>()
+            .HasIndex("ClusterNodeTypeAggregationId")
+            .HasFilter("[IsDeleted] = 0")
+            .IncludeProperties("AccountingId");
+
+        modelBuilder.Entity<ClusterProjectCredential>()
+            .HasIndex(cpc => cpc.ClusterProjectId)
+            .HasFilter("[IsDeleted] = 0")
+            .IncludeProperties(cpc => cpc.ClusterAuthenticationCredentialsId);
+
+        modelBuilder.Entity<ClusterProject>()
+            .HasIndex(cp => cp.ClusterId)
+            .HasFilter("[IsDeleted] = 0")
+            .IncludeProperties(cp => cp.ProjectId);
+
+        modelBuilder.Entity<ClusterProject>()
+            .HasIndex(cp => cp.ProjectId)
+            .HasFilter("[IsDeleted] = 0")
+            .IncludeProperties(cp => cp.ClusterId);
         
         modelBuilder.Entity<SessionCode>()
             .HasIndex(s => s.UniqueCode)