+++ DataAccessTier/Repository/JobManagement/JobSpecificationRepository.cs
@@ -1,7 +1,8 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Linq;
 using HEAppE.DataAccessTier.IRepository.JobManagement;
 using HEAppE.DomainObjects.JobManagement;
+using Microsoft.EntityFrameworkCore;
 
 namespace HEAppE.DataAccessTier.Repository.JobManagement;
 
@@ -22,6 +23,14 @@ internal class JobSpecificationRepository : GenericRepository<JobSpecification>,
     {
         return _dbSet.Where(js => js.FileTransferMethodId == fileTransferMethodId)
             .ToList();
+    }
+
+    public JobSpecification GetByIdWithTasksAndSubmitter(long id)
+    {
+        return _dbSet
+            .Include(js => js.Tasks)
+            .Include(js => js.Submitter)
+            .FirstOrDefault(js => js.Id == id);
     }
 
     #endregion