+++ DataAccessTier/Repository/UserAndLimitationManagement/AdaptorUserRepository.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Linq;
 using HEAppE.DataAccessTier.IRepository.UserAndLimitationManagement;
 using HEAppE.DomainObjects.UserAndLimitationManagement;
@@ -22,11 +22,10 @@ internal class AdaptorUserRepository : GenericRepository<AdaptorUser>, IAdaptorU
     public AdaptorUser GetByName(string username)
     {
         return _dbSet
-            // 1. Větev: Načtení rolí (bez ContainedRoleTypes, ty se načtou automaticky s rolí)
+            .AsSplitQuery()
             .Include(u => u.AdaptorUserUserGroupRoles)
             .ThenInclude(ugr => ugr.AdaptorUserRole)
-        
-            // 2. Větev: Načtení skupin a projektů
+
             .Include(u => u.AdaptorUserUserGroupRoles)
             .ThenInclude(ugr => ugr.AdaptorUserGroup)
             .ThenInclude(ug => ug.Project)
@@ -39,6 +38,7 @@ internal class AdaptorUserRepository : GenericRepository<AdaptorUser>, IAdaptorU
     public AdaptorUser GetByApiKey(string apiKey)
     {
         return _dbSet
+            .AsSplitQuery()
             .Include(u => u.AdaptorUserUserGroupRoles)
             .ThenInclude(ugr => ugr.AdaptorUserRole)
             .Include(u => u.AdaptorUserUserGroupRoles)
@@ -53,6 +53,7 @@ internal class AdaptorUserRepository : GenericRepository<AdaptorUser>, IAdaptorU
     public override AdaptorUser GetById(long id)
     {
         return _dbSet
+            .AsSplitQuery()
             .Include(u => u.AdaptorUserUserGroupRoles)
             .ThenInclude(ugr => ugr.AdaptorUserRole)
             .Include(u => u.AdaptorUserUserGroupRoles)